bout_runners update and fix
* Cleaned up examples in example folder
* Added new example
* Moved `restart_from_func` to `pre_and_post_processing`
