+++ examples/bout_runners_example/13-restart_w_add_noise.py
@@ -1,21 +1,13 @@
 #!/usr/bin/env python
 
-"""Driver which restarts a scan, given a restart function"""
+"""Driver which resizes the grid after restart"""
 
 from pre_and_post_processing.post_processing_show_the_data import show_the_data
-from pre_and_post_processing.restart_from_func import restart_from_func
 from bout_runners import basic_runner
 
-scan = (("cst", "D_perp", (1.0,5.5)),\
-        ("cst", "D_par",  (1.5,2.5))
-       )
-
-# Given that the runs has already been performed
-only_post_process = False
-
-# Initial runs
+# Initial run
 # ===========================================================================
-init_run = basic_runner(additional = scan)
+init_run = basic_runner(nz = 16)
 
 dmp_folder, _ =\
         init_run.execute_runs(\
@@ -30,22 +22,20 @@ dmp_folder, _ =\
                               y = slice(0,None),\
                               z = slice(0,None)\
                              )
-
-one_of_the_restart_paths_in_scan = dmp_folder[0]
 # ===========================================================================
 
 
-# Restart the scan
+# Restart the run after resizing the grid
 # ===========================================================================
-if only_post_process:
-    restart = None
-else:
-    restart = "overwrite"
-
-restart_run = basic_runner(nout         = 5                ,\
-                           restart      = restart          ,\
-                           restart_from = restart_from_func,\
-                           additional   = scan             ,\
+restart_run = basic_runner(restart      = "overwrite"  ,\
+                           restart_from = dmp_folder[0],\
+                           nx           = 22           ,\
+                           ny           = 22           ,\
+                           nz           = 16           ,\
+                           # NOTE: This amount if noise is large
+                           #       relative to the backgroun, and is
+                           #       just added for illustrative purposes
+                           add_noise    = {"n": 5e-4}  ,\
                            )
 
 restart_run.execute_runs(\
@@ -58,11 +48,6 @@ restart_run.execute_runs(\
                          t = slice(0,None),\
                          x = 1,\
                          y = slice(0,None),\
-                         z = slice(0,None),\
-                         # Below are the kwargs given to the
-                         # restart_from_func
-                         one_of_the_restart_paths_in_scan =\
-                         one_of_the_restart_paths_in_scan,\
-                         scan_parameters = ["D_perp", "D_par"],\
+                         z = slice(0,None)\
                         )
 # ===========================================================================
